move find_scale_and_prefix to non mpl plotting
