+++ qcodes/plotting/__init__.py
@@ -9,9 +9,8 @@ from .matplotlib_helpers import (
     apply_auto_color_scale,
     apply_color_scale_limits,
     auto_color_scale_from_config,
-    find_scale_and_prefix,
 )
-from .plotting import auto_range_iqr
+from .plotting import auto_range_iqr, find_scale_and_prefix
 
 __all__ = [
     "apply_auto_color_scale",
